+++ crates/simulation/src/tourism/systems.rs
@@ -1,0 +1,42 @@
+use bevy::prelude::*;
+
+use crate::services::ServiceBuilding;
+use crate::stats::CityStats;
+use crate::weather::Weather;
+
+use super::{tourism_seasonal_modifier, Tourism};
+
+pub fn update_tourism(
+    clock: Res<crate::time_of_day::GameClock>,
+    mut tourism: ResMut<Tourism>,
+    services: Query<&ServiceBuilding>,
+    stats: Res<CityStats>,
+    weather: Res<Weather>,
+) {
+    // Update monthly
+    if clock.day <= tourism.last_update_day + 30 {
+        return;
+    }
+    tourism.last_update_day = clock.day;
+
+    // Calculate attractiveness from landmarks and entertainment
+    let mut total_draw = 0u32;
+    for service in &services {
+        total_draw += Tourism::tourism_draw(service.service_type);
+    }
+
+    // Attractiveness scales with city size and landmarks
+    let pop_factor = (stats.population as f32 / 10000.0).min(5.0);
+    tourism.attractiveness = (total_draw as f32 * 0.1 + pop_factor * 10.0).min(100.0);
+
+    // Visitors based on attractiveness, boosted by airport and weather/season modifiers
+    let base_visitors = (tourism.attractiveness * 50.0) as u32;
+    let season_weather_modifier = tourism_seasonal_modifier(weather.season, &weather);
+    tourism.monthly_visitors =
+        (base_visitors as f32 * tourism.airport_multiplier * season_weather_modifier) as u32;
+
+    // Tourism income: visitors spend money at commercial buildings
+    // Airport multiplier also boosts per-visitor spending (international travelers spend more)
+    let spending_per_visitor = 2.0 * tourism.airport_multiplier as f64;
+    tourism.monthly_tourism_income = tourism.monthly_visitors as f64 * spending_per_visitor;
+}
